create simple fe cli interface
