+++ crates/fe/Cargo.toml
@@ -14,7 +14,7 @@ version = "0.18.0-alpha"
 solc-backend = ["fe-driver/solc-backend"]
 
 [dependencies]
-clap = "2.33.3"
+clap = {version="3.1.18", features = ["derive"]}
 walkdir = "2"
 indexmap = "1.6.2"
